+++ opencti-platform/opencti-graphql/src/database/draft-utils.js
@@ -5,7 +5,6 @@ import { READ_INDEX_DRAFT_OBJECTS, toBase64, UPDATE_OPERATION_ADD, UPDATE_OPERAT
 import { DRAFT_OPERATION_CREATE, DRAFT_OPERATION_DELETE, DRAFT_OPERATION_DELETE_LINKED, DRAFT_OPERATION_UPDATE } from '../modules/draftWorkspace/draftOperations';
 import { EditOperation } from '../generated/graphql';
 import { getFileContent, loadFile } from './file-storage';
-import { INPUT_MARKINGS } from '../schema/general';
 
 export const getDraftFilePrefix = (draftId) => {
   return `draft/${draftId}/`;
@@ -58,13 +57,13 @@ export const isDraftSupportedEntity = (element) => {
   return !isInternalObject(element.entity_type) && !isInternalRelationship(element.entity_type);
 };
 
-const filesKey = 'x_opencti_files';
+export const FILES_UPDATE_KEY = 'x_opencti_files';
 // Transform a raw update patched stored in a draft_updates_patch to a list of reverse field patch inputs
 export const buildReverseUpdateFieldPatch = (rawUpdatePatch) => {
   const resulReverseFieldPatch = [];
   if (rawUpdatePatch) {
     const parsedUpdatePatch = JSON.parse(rawUpdatePatch);
-    const updatePatchKeys = Object.keys(parsedUpdatePatch).filter((k) => k !== filesKey);
+    const updatePatchKeys = Object.keys(parsedUpdatePatch).filter((k) => k !== FILES_UPDATE_KEY);
     for (let i = 0; i < updatePatchKeys.length; i += 1) {
       const currentKey = updatePatchKeys[i];
       const currentValues = parsedUpdatePatch[currentKey];
@@ -79,11 +78,11 @@ export const buildReverseUpdateFieldPatch = (rawUpdatePatch) => {
 };
 
 // Transform a raw update patched stored in a draft_updates_patch to a list of field patch inputs
-export const buildUpdateFieldPatch = async (context, user, rawUpdatePatch) => {
+export const buildUpdateFieldPatch = (rawUpdatePatch) => {
   const resultFieldPatch = [];
   if (rawUpdatePatch) {
     const parsedUpdatePatch = JSON.parse(rawUpdatePatch);
-    const updatePatchKeys = Object.keys(parsedUpdatePatch).filter((k) => k !== filesKey);
+    const updatePatchKeys = Object.keys(parsedUpdatePatch);
     for (let i = 0; i < updatePatchKeys.length; i += 1) {
       const currentKey = updatePatchKeys[i];
       const currentValues = parsedUpdatePatch[currentKey];
@@ -101,28 +100,6 @@ export const buildUpdateFieldPatch = async (context, user, rawUpdatePatch) => {
             resultFieldPatch.push(removeInput);
           }
         }
-      }
-    }
-    if (updatePatchKeys.find((k) => k === filesKey)) {
-      const fileValues = parsedUpdatePatch[filesKey];
-      if (fileValues) {
-        const fileIds = fileValues.added_value;
-        const loadedFileValues = [];
-        for (let i = 0; i < fileIds.length; i += 0) {
-          const currentFileId = fileIds[i];
-          const currentFile = await loadFile(context, user, currentFileId);
-          const currentFileContent = toBase64(await getFileContent(currentFileId));
-          const currentFileObject = {
-            name: currentFile.name,
-            data: currentFileContent,
-            version: currentFile.metaData.version,
-            mime_type: currentFile.metaData.mime_type,
-            object_marking_refs: currentFile.metaData.file_markings ?? [],
-          };
-          loadedFileValues.push(currentFileObject);
-        }
-        const addInput = { key: filesKey, value: loadedFileValues, operation: EditOperation.Add };
-        resultFieldPatch.push(addInput);
       }
     }
   }
